// src/app/hotel/types.ts
export interface HotelSearchResultHotelInfo {
  id: ID
  name: string
  slug: string
  zip_code: string | null
  address: string
  destination: string
  country_code: string
  location: number[]
  phone: string | null
  email: string | null
  currency: string
  meal_type: string
  nearby: string[] | [] | null
  checkin_from: string
  checkout_to: string
  fax: string | null
  nr_rooms: null | number
  stars: number
  availability_score: null
  max_free_child_age: null | number
  min_free_child_age: null | number
  images: {
    category: number
    tag: string
    original: null
    large: string
    small: null | string
    mid: null | string
    default: boolean
    priority: number
  }[]
  themes: number[] | []
  facilities: number[] | []
  facility_scopes: null
  tripAdvisor: null
  price: ServicePriceType
  descriptions: null
  year_built: null | number | string
  nr_restaurants: null | number
  nr_bars: null | number
  nr_halls: null | number
  last_update: string
  video_list: null | string[]
  web_site: '' | string | null
  deleted: boolean
  destination_slug: string | null
  old_destination_slug: null
  rating: number
  listing_rate: number
  sales_rate: number
  destination_id: ID
  destination_map: number[]
  search_rate: number
  reviews: null
  nearby_restaurants: null
  nearby_info: null | string
  comment_info: null | {
    comments: null | string
    totalComments: number
    averageScore: number
  }
  documents: null
  food_drinks: null
}
// ...
export interface HotelCampaignsResponse {
  id: ID
  title: string
  typeId: ID
  collectionId: ID
  point: string
  params: {
    sort_description: {
      value: string
    }
    description: {
      value: string
    }
    btn_text: {
      value: string
    }
    link: {
      value: string
    }
    image: {
      value: string
    }
    svg: {
      value: string
    }
    view_country: {
      value: string
    }
    slug: {
      value: string
    }
  }
  ordering: number
  language: string
  active: boolean
}